Time: 0 ms (100%), Space: 103.5 MB (96.89%) - LeetHub

+++ 3689-maximum-total-subarray-value-i/3689-maximum-total-subarray-value-i.cpp
@@ -4,6 +4,6 @@ public:
         int maxi = *max_element(nums.begin() , nums.end()) ;
         int mini = *min_element(nums.begin() , nums.end()) ;
 
-        return (long long)(k*(long long)(maxi-mini)) ;
+        return (k*(long long)(maxi-mini)) ;
     }
 };
